Refactor configuration management in Bitloops CLI to support new store backends. Updated documentation to reflect changes in configuration structure, emphasizing repository-level settings and semantic features. Introduced new modules for agent hooks and store backends, enhancing initialization processes and telemetry consent handling. Improved error handling and streamlined command execution for better user experience.
