+++ bitloops_cli/src/commands/explain_test.rs
@@ -56,16 +56,13 @@ fn make_checkpoint_commit(
 }
 
 fn checkpoint_sqlite_path(repo_root: &std::path::Path) -> PathBuf {
-    let cfg =
-        crate::devql_config::resolve_devql_backend_config().expect("resolve devql backend config");
+    let cfg = crate::store_config::resolve_store_backend_config_for_repo(repo_root)
+        .expect("resolve backend config");
     if let Some(path) = cfg.relational.sqlite_path.as_deref() {
-        crate::devql_config::resolve_sqlite_db_path(Some(path))
+        crate::store_config::resolve_sqlite_db_path_for_repo(repo_root, Some(path))
             .expect("resolve configured sqlite path")
     } else {
-        repo_root
-            .join(crate::engine::paths::BITLOOPS_DIR)
-            .join("devql")
-            .join("relational.db")
+        crate::engine::paths::default_relational_db_path(repo_root)
     }
 }
