+++ bitloops_cli/src/engine/db/mod.rs
@@ -3,7 +3,7 @@ pub mod sqlite;
 
 use anyhow::{Context, Result};
 
-use crate::devql_config::DevqlBackendConfig;
+use crate::store_config::StoreBackendConfig;
 
 pub use postgres::PostgresSyncConnection;
 pub use sqlite::SqliteConnectionPool;
@@ -15,12 +15,12 @@ pub struct CheckpointDbConnections {
 }
 
 impl CheckpointDbConnections {
-    pub fn connect_from_devql_config(cfg: &DevqlBackendConfig) -> Result<Self> {
+    pub fn connect_from_store_config(cfg: &StoreBackendConfig) -> Result<Self> {
         let sqlite_path = cfg
             .relational
             .resolve_sqlite_db_path()
             .context("resolving SQLite path for checkpoint storage")?;
-        let sqlite = SqliteConnectionPool::connect(sqlite_path)?;
+        let sqlite = SqliteConnectionPool::connect_existing(sqlite_path)?;
         let postgres = cfg
             .relational
             .postgres_dsn
@@ -51,7 +51,7 @@ impl CheckpointDbConnections {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::devql_config::resolve_devql_backend_config_for_tests;
+    use crate::store_config::{StoreFileConfig, resolve_store_backend_config_for_tests};
     use anyhow::Result;
     use tempfile::TempDir;
 
@@ -59,11 +59,21 @@ mod tests {
     fn checkpoint_db_connections_initialise_sqlite_schema_even_without_postgres() -> Result<()> {
         let temp = TempDir::new().context("creating temp dir for sqlite")?;
         let sqlite_path = temp.path().join("db").join("relational.sqlite");
-        let sqlite_path_env = sqlite_path.to_string_lossy().to_string();
-        let env = [("BITLOOPS_DEVQL_SQLITE_PATH", sqlite_path_env.as_str())];
+        std::fs::create_dir_all(
+            sqlite_path
+                .parent()
+                .expect("sqlite path should have parent directory"),
+        )
+        .context("creating sqlite parent directory for test")?;
+        let _ = rusqlite::Connection::open(&sqlite_path)
+            .context("creating sqlite file for connect_existing test")?;
+        let file_cfg = StoreFileConfig {
+            sqlite_path: Some(sqlite_path.to_string_lossy().to_string()),
+            ..Default::default()
+        };
 
-        let cfg = resolve_devql_backend_config_for_tests(Default::default(), &env)?;
-        let connections = CheckpointDbConnections::connect_from_devql_config(&cfg)?;
+        let cfg = resolve_store_backend_config_for_tests(file_cfg)?;
+        let connections = CheckpointDbConnections::connect_from_store_config(&cfg)?;
         connections.initialise_checkpoint_schema()?;
 
         let sessions_table_exists = connections.sqlite().with_connection(|conn| {
@@ -93,17 +103,17 @@ mod tests {
     fn checkpoint_db_connections_enable_optional_postgres_when_dsn_present() -> Result<()> {
         let temp = TempDir::new().context("creating temp dir for sqlite")?;
         let sqlite_path = temp.path().join("relational.sqlite");
-        let sqlite_path_env = sqlite_path.to_string_lossy().to_string();
-        let env = [
-            ("BITLOOPS_DEVQL_SQLITE_PATH", sqlite_path_env.as_str()),
-            (
-                "BITLOOPS_DEVQL_PG_DSN",
-                "postgres://bitloops:bitloops@localhost:5432/bitloops",
-            ),
-        ];
+        let _ = rusqlite::Connection::open(&sqlite_path)
+            .context("creating sqlite file for connect_existing test")?;
+        let file_cfg = StoreFileConfig {
+            relational_provider: Some("postgres".to_string()),
+            sqlite_path: Some(sqlite_path.to_string_lossy().to_string()),
+            pg_dsn: Some("postgres://bitloops:bitloops@localhost:5432/bitloops".to_string()),
+            ..Default::default()
+        };
 
-        let cfg = resolve_devql_backend_config_for_tests(Default::default(), &env)?;
-        let connections = CheckpointDbConnections::connect_from_devql_config(&cfg)?;
+        let cfg = resolve_store_backend_config_for_tests(file_cfg)?;
+        let connections = CheckpointDbConnections::connect_from_store_config(&cfg)?;
 
         assert!(connections.postgres().is_some());
         Ok(())
