+++ bitloops_cli/src/engine/paths/mod.rs
@@ -10,6 +10,12 @@ pub const BITLOOPS_DIR: &str = ".bitloops";
 pub const BITLOOPS_TMP_DIR: &str = ".bitloops/tmp";
 // Legacy compatibility path used by git-backed checkpoint metadata.
 pub const BITLOOPS_METADATA_DIR: &str = ".bitloops/metadata";
+pub const BITLOOPS_STORES_DIR: &str = ".bitloops/stores";
+pub const BITLOOPS_RELATIONAL_STORE_DIR: &str = ".bitloops/stores/relational";
+pub const BITLOOPS_EVENT_STORE_DIR: &str = ".bitloops/stores/event";
+pub const BITLOOPS_BLOB_STORE_DIR: &str = ".bitloops/stores/blob";
+pub const RELATIONAL_DB_FILE_NAME: &str = "relational.db";
+pub const EVENTS_DB_FILE_NAME: &str = "events.duckdb";
 
 // Metadata file names.
 pub const CONTEXT_FILE_NAME: &str = "context.md";
@@ -233,6 +239,22 @@ pub fn abs_path(path: &str) -> Result<PathBuf> {
 /// Returns `.bitloops/metadata/<session_id>`.
 pub fn session_metadata_dir_from_session_id(session_id: &str) -> String {
     format!("{BITLOOPS_METADATA_DIR}/{session_id}")
+}
+
+pub fn default_relational_db_path(repo_root: &Path) -> PathBuf {
+    repo_root
+        .join(BITLOOPS_RELATIONAL_STORE_DIR)
+        .join(RELATIONAL_DB_FILE_NAME)
+}
+
+pub fn default_events_db_path(repo_root: &Path) -> PathBuf {
+    repo_root
+        .join(BITLOOPS_EVENT_STORE_DIR)
+        .join(EVENTS_DB_FILE_NAME)
+}
+
+pub fn default_blob_store_path(repo_root: &Path) -> PathBuf {
+    repo_root.join(BITLOOPS_BLOB_STORE_DIR)
 }
 
 /// Attempts to extract a session ID from a transcript path.
