+++ bitloops_cli/src/server/dashboard/mod.rs
@@ -6,11 +6,11 @@ mod handlers;
 mod router;
 
 use crate::branding::{BITLOOPS_PURPLE_HEX, bitloops_wordmark, color_hex};
-use crate::devql_config::dashboard_use_bitloops_local;
 use crate::engine::paths;
 use crate::engine::strategy::manual_commit::{
     CommittedInfo, list_committed, read_commit_checkpoint_mappings, read_committed_info, run_git,
 };
+use crate::store_config::dashboard_use_bitloops_local;
 use anyhow::{Context, Result, anyhow, bail};
 use std::collections::{HashMap, HashSet};
 use std::env;
