+++ bitloops_cli/tests/e2e_scenario_groups.rs
@@ -226,17 +226,17 @@ fn init_and_enable(repo: &Path) {
 
     // Pre-track repo-local storage paths so stash -u scenarios don't conflict
     // on untracked DevQL/Blob files between prompts.
-    let devql_dir = repo.join(".bitloops/devql");
-    fs::create_dir_all(&devql_dir).expect("create .bitloops/devql");
-    let sqlite_path = devql_dir.join("relational.db");
+    let relational_dir = repo.join(".bitloops/stores/relational");
+    fs::create_dir_all(&relational_dir).expect("create .bitloops/stores/relational");
+    let sqlite_path = relational_dir.join("relational.db");
     if !sqlite_path.exists() {
-        fs::File::create(&sqlite_path).expect("create .bitloops/devql/relational.db");
+        fs::File::create(&sqlite_path).expect("create .bitloops/stores/relational/relational.db");
     }
-    let blobs_dir = repo.join(".bitloops/blobs");
-    fs::create_dir_all(&blobs_dir).expect("create .bitloops/blobs");
+    let blobs_dir = repo.join(".bitloops/stores/blob");
+    fs::create_dir_all(&blobs_dir).expect("create .bitloops/stores/blob");
     let keep = blobs_dir.join(".gitkeep");
     if !keep.exists() {
-        fs::write(&keep, "").expect("create .bitloops/blobs/.gitkeep");
+        fs::write(&keep, "").expect("create .bitloops/stores/blob/.gitkeep");
     }
 
     // Keep infrastructure files tracked so stash/pop scenarios do not conflict
@@ -2122,7 +2122,11 @@ fn cli_1159_scenario6_stash_second_prompt_unstash_commit_all() {
     // Reconcile tracked SQLite changes so stash pop can restore stashed untracked files.
     run_git_expect_success(
         dir.path(),
-        &["restore", "--worktree", ".bitloops/devql/relational.db"],
+        &[
+            "restore",
+            "--worktree",
+            ".bitloops/stores/relational/relational.db",
+        ],
         "restore relational.db before stash pop",
     );
     run_git_expect_success(dir.path(), &["stash", "pop"], "stash pop combo_b/combo_c");
